Refactor: Move transaction window import to gui module
